Add functional tests
The functional tests are themselves Gauge tests.

The scaffolding for the tests has been lifted and shifted from the
[gauge-tests][1] repository. When adding more Spectacle functional tests
definitely browse the `gauge-tests` repository for inspiration and
ideas.

This commit only adds a small number of functional tests with limited
coverage of Spectacle's functionality, but as these are the first
functional tests for Spectacle there is still significant
value from a regression prevention perspective.  Further tests can be
added in future commits if deemed useful.

The functional tests run on every push and pull request, triggered by
a GitHub Action.

[1]: https://github.com/getgauge/gauge-tests

Signed-off-by: John Boyes <154404+johnboyes@users.noreply.github.com>